move RSS and Cube to data_container module and refactoring
