+++ src/pykoala/cubing.py
@@ -15,7 +15,7 @@ from astropy.coordinates import SkyCoord
 # KOALA packages
 # =============================================================================
 from pykoala import ancillary
-from pykoala.data_container import SpectraContainer
+from pykoala.data_container import Cube
 from pykoala.plotting import qc_plot
 from pykoala import __version__, vprint
 from scipy.special import erf
@@ -545,248 +545,6 @@ def build_hdul(intensity, variance, primary_header_info=None, wcs=None):
     fits.ImageHDU(name='VARIANCE', data=variance, header=data_header)])
 
     return hdul
-
-
-
-# =============================================================================
-# Cube class
-# =============================================================================
-
-class Cube(SpectraContainer):
-    """This class represent a collection of Raw Stacked Spectra (RSS) interpolated over a 2D spatial grid.
-    
-    parent_rss
-    rss_mask
-    intensity
-    variance
-    intensity
-    variance
-    wavelength
-    info
-
-    """
-    n_wavelength = None
-    n_cols = None
-    n_rows = None
-    x_size_arcsec = None
-    y_size_arcsec = None
-    default_hdul_extensions_map = {"INTENSITY": "INTENSITY",
-                                   "VARIANCE": "VARIANCE"}
-    
-    def __init__(self, hdul=None, hdul_extensions_map=None, **kwargs):
-
-        self.hdul = hdul
-
-        if hdul_extensions_map is not None:
-            self.hdul_extensions_map = hdul_extensions_map
-        else:
-            self.hdul_extensions_map = self.default_hdul_extensions_map
-
-        if "logger" not in kwargs:
-            kwargs['logger'] = "pykoala.cube"
-        super().__init__(intensity=self.intensity,
-                         variance=self.variance,
-                         **kwargs)
-        self.get_wcs_from_header()
-        self.parse_info_from_header()
-        self.n_wavelength, self.n_rows, self.n_cols = self.intensity.shape
-        self.get_wavelength()
-
-    @property
-    def hdul(self):
-        return self._hdul
-
-    @hdul.setter
-    def hdul(self, hdul):
-        assert isinstance(hdul, fits.HDUList)
-        self._hdul = hdul
-
-    @property
-    def intensity(self):
-        return self.hdul[self.hdul_extensions_map['INTENSITY']].data
-    
-    @intensity.setter
-    def intensity(self, intensity_corr):
-        self.vprint("[Cube] Updating HDUL INTENSITY")
-        self.hdul[self.hdul_extensions_map['INTENSITY']].data = intensity_corr
-
-    @property
-    def variance(self):
-        return self.hdul[self.hdul_extensions_map['VARIANCE']].data
-
-    @variance.setter
-    def variance(self, variance_corr):
-        self.vprint("[Cube] Updating HDUL variance")
-        self.hdul[self.hdul_extensions_map['VARIANCE']].data = variance_corr
-
-    @property
-    def rss_intensity(self):
-        return np.reshape(self.intensity, (self.intensity.shape[0], self.intensity.shape[1]*self.intensity.shape[2])).T
-
-    @rss_intensity.setter   
-    def rss_intensity(self, value):
-        self.intensity = value.T.reshape(self.intensity.shape)
-
-    @property
-    def rss_variance(self):
-        return np.reshape(self.variance, (self.variance.shape[0], self.variance.shape[1]*self.variance.shape[2])).T
-
-    @rss_variance.setter   
-    def rss_variance(self, value):
-        self.variance = value.T.reshape(self.variance.shape)
-
-    @classmethod
-    def from_fits(cls, path, hdul_extension_map=None, **kwargs):
-        """Make an instance of a Cube using an input path to a FITS file.
-        
-        Parameters
-        ----------
-        - path: str
-            Path to the FITS file. This file must be compliant with the pykoala
-            standards.
-        - hdul_extension_map: dict
-            Dictionary containing the mapping to access the extensions that
-            contain the intensity, and variance data
-            (e.g. {'INTENSITY': 1, 'VARIANCE': 'var'}).
-        - kwargs:
-            Arguments passed to the Cube class (see Cube documentation)
-        
-        Returns
-        -------
-        - cube: Cube
-            An instance of a `pykoala.cubing.Cube`.
-        """
-        with fits.open(path) as hdul:
-            return cls(hdul, hdul_extension_map=hdul_extension_map, **kwargs)
-
-    def parse_info_from_header(self):
-        """Look into the primary header for pykoala information."""
-        self.vprint("[Cube] Looking for information in the primary header")
-        # TODO
-        #self.info = {}
-        #self.fill_info()
-        self.history.load_from_header(self.hdul[0].header)
-
-    def load_hdul(self, path_to_file):
-        self.hdul = fits.open(path_to_file)
-        pass
-
-    def close_hdul(self):
-        if self.hdul is not None:
-            self.vprint(f"[Cube] Closing HDUL")
-            self.hdul.close()
-
-    def get_wcs_from_header(self):
-        """Create a WCS from HDUL header."""
-        self.wcs = WCS(self.hdul[self.hdul_extensions_map['INTENSITY']].header)
-
-    def get_wavelength(self):
-        self.vprint("[Cube] Constructing wavelength array")
-        self.wavelength = self.wcs.spectral.array_index_to_world(
-            np.arange(self.n_wavelength)).to('angstrom').value
-        
-    def get_centre_of_mass(self, wavelength_step=1, stat=np.median, power=1.0):
-        """Compute the center of mass of the data cube."""
-        x = np.arange(0, self.n_cols, 1)
-        y = np.arange(0, self.n_rows, 1)
-        x_com = np.empty(self.n_wavelength)
-        y_com = np.empty(self.n_wavelength)
-        for wave_range in range(0, self.n_wavelength, wavelength_step):
-            x_com[wave_range: wave_range + wavelength_step] = np.nansum(
-                self.intensity[wave_range: wave_range + wavelength_step]**power * x[np.newaxis, np.newaxis, :],
-                axis=(1, 2)) / np.nansum(self.intensity[wave_range: wave_range + wavelength_step]**power, axis=(1, 2))
-            x_com[wave_range: wave_range + wavelength_step] = stat(x_com[wave_range: wave_range + wavelength_step])
-            y_com[wave_range: wave_range + wavelength_step] = np.nansum(
-                self.intensity[wave_range: wave_range + wavelength_step]**power * y[np.newaxis, :, np.newaxis],
-                axis=(1, 2)) / np.nansum(self.intensity[wave_range: wave_range + wavelength_step]**power, axis=(1, 2))
-            y_com[wave_range: wave_range + wavelength_step] = stat(y_com[wave_range: wave_range + wavelength_step])
-        return x_com, y_com
-
-    def get_integrated_light_frac(self, frac=0.5):
-        """Compute the integrated spectra that accounts for a given fraction of the total intensity."""
-        collapsed_intensity = np.nansum(self.intensity, axis=0)
-        sort_intensity = np.sort(collapsed_intensity, axis=(0, 1))
-        # Sort from highes to lowest luminosity
-        sort_intensity = np.flip(sort_intensity, axis=(0, 1))
-        cumulative_intensity = np.cumsum(sort_intensity, axis=(0, 1))
-        cumulative_intensity /= np.nanmax(cumulative_intensity)
-        pos = np.searchsorted(cumulative_intensity, frac)
-        return cumulative_intensity[pos]
-
-    def get_white_image(self, wave_range=None, s_clip=3.0, frequency_density=False):
-        """Create a white image."""
-        if wave_range is not None and self.wavelength is not None:
-            wave_mask = (self.wavelength >= wave_range[0]) & (self.wavelength <= wave_range[1])
-        else:
-            wave_mask = np.ones(self.wavelength.size, dtype=bool)
-        
-        if s_clip is not None:
-            std_dev = ancillary.std_from_mad(self.intensity[wave_mask], axis=0)
-            median = np.nanmedian(self.intensity[wave_mask], axis=0)
-
-            weights = (
-                (self.intensity[wave_mask] <= median[np.newaxis] + s_clip * std_dev[np.newaxis])
-                & (self.intensity[wave_mask] >= median[np.newaxis] - s_clip * std_dev[np.newaxis]))
-        else:
-            weights = np.ones_like(self.intensity[wave_mask])
-
-        if frequency_density:
-            freq_trans = self.wavelength**2 / 3e18
-        else:
-            freq_trans = np.ones_like(self.wavelength)
-
-        white_image = np.nansum(
-            self.intensity[wave_mask] * freq_trans[wave_mask, np.newaxis, np.newaxis] * weights, axis=0
-            ) / np.nansum(weights, axis=0)
-        return white_image
-
-    def to_fits(self, fname=None, primary_hdr_kw=None):
-        """Save the Cube into a FITS file."""
-        if fname is None:
-            fname = 'cube_{}.fits.gz'.format(
-                datetime.now().strftime("%d_%m_%Y_%H_%M_%S"))
-        if primary_hdr_kw is None:
-            primary_hdr_kw = {}
-
-        # Create the PrimaryHDU with WCS information 
-        primary = fits.PrimaryHDU()
-        for key, val in primary_hdr_kw.items():
-            primary.header[key] = val
-        
-        
-        # Include cubing information
-        primary.header['pykoala0'] = __version__, "PyKOALA version"
-        primary.header['pykoala1'] = datetime.now().strftime(
-            "%d_%m_%Y_%H_%M_%S"), "creation date / last change"
-
-        # Fill the header with the log information
-        primary.header = self.history.dump_to_header(primary.header)
-
-        # Create a list of HDU
-        hdu_list = [primary]
-        # Change headers for variance and INTENSITY
-        hdu_list.append(fits.ImageHDU(
-            data=self.intensity, name='INTENSITY', header=self.wcs.to_header()))
-        hdu_list.append(fits.ImageHDU(
-            data=self.variance, name='VARIANCE', header=hdu_list[-1].header))
-        # Store the mask information
-        hdu_list.append(self.mask.dump_to_hdu())
-        # Save fits
-        hdul = fits.HDUList(hdu_list)
-        hdul.writeto(fname, overwrite=True)
-        hdul.close()
-        self.vprint("[Cube] Cube saved at:\n {}".format(fname))
-
-    def update_coordinates(self, new_coords=None, offset=None):
-        """Update the celestial coordinates of the Cube"""
-        updated_wcs = ancillary.update_wcs_coords(self.wcs.celestial,
-                                               ra_dec_val=new_coords,
-                                               ra_dec_offset=offset)
-        # Update only the celestial axes
-        self.vprint("Previous CRVAL: ", self.wcs.celestial.wcs.crval,
-              "\nNew CRVAL: ", updated_wcs.wcs.crval)
-        self.wcs.wcs.crval[:-1] = updated_wcs.wcs.crval
-        self.history('update_coords', "Offset-coords updated")
 
 def make_white_image_from_array(data_array, wavelength=None, **args):
     """Create a white image from a 3D data array."""
