+++ src/pykoala/spectra/nebular_emission_line_list.txt
@@ -1,0 +1,236 @@
+
+#λ (AA), Ion
+770.409, NeVIII
+780.324, NeVIII
+937.814, Lye
+949.742, Lyd
+977.030, CIII
+989.790, NIII
+991.514, NIII
+991.579, NIII
+1025.722, Lyb
+1031.912, OVI
+1037.613, OVI
+1066.660, ArI
+1215.670, Lya
+1238.821, NV
+1242.804, NV
+1260.422, SiII
+1264.730, SiII
+1302.168, OI
+1334.532, CII
+1335.708, CII
+1393.755, SiIV
+1397.232, OIV]
+1399.780, OIV]
+1402.770, SiIV
+1486.496, NIV]
+1548.187, CIV
+1550.772, CIV
+1640.420, HeII
+1660.809, OIII]
+1666.150, OIII]
+1746.823, NIII]
+1748.656, NIII]
+1854.716, AlIII
+1862.790, AlIII
+1892.030, SiIII]
+1908.734, CIII]
+2142.780, NII]
+2320.951, [OIII]
+2323.500, CII]
+2324.690, CII]
+2648.710, [FeXI]
+2733.289, HeII
+2782.700, [MgV]
+2795.528, MgII]
+2802.705, MgII]
+2829.360, [FeIV]
+2835.740, [FeIV]
+2853.670, [ArIV]
+2868.210, [ArIV]
+2928.000, [MgV]
+2945.106, HeI
+3132.794, OIII
+3187.745, HeI
+3203.100, HeII
+3312.329, OIII
+3345.821, [NeV]
+3425.881, [NeV]
+3444.052, OIII
+3466.497, [NI]
+3466.543, [NI]
+3487.727, HeI
+3586.320, [FeVII]
+3662.500, [FeVI]
+3686.831, H19
+3691.551, H18
+3697.157, H17
+3703.859, H16
+3711.977, H15
+3721.945, H14
+3726.032, [OII]
+3728.815, [OII]
+3734.369, H13
+3750.158, H12
+3758.920, [FeVII]
+3770.637, H11
+3797.904, H10
+3835.391, H9
+3839.270, [FeV]
+3868.760, [NeIII]
+3888.647, HeI
+3889.064, H8
+3891.280, [FeV]
+3911.330, [FeV]
+3967.470, [NeIII]
+3970.079, He
+4026.190, HeI20.9
+4068.600, [SII]
+4071.240, [FeV]
+4076.349, [SII]
+4101.742, Hd
+4143.761, HeI
+4178.862, FeII
+4180.600, [FeV]
+4233.172, FeII
+4227.190, [FeV]
+4287.394, [FeII]
+4303.176, FeII
+4317.139, OII
+4340.471, Hg
+4363.210, [OIII]
+4412.300, [ArXIV]
+4414.899, OII
+4416.830, FeII
+4452.098, [FeII]
+4471.479, HeI
+4489.183, FeII
+4491.405, FeII
+4510.910, NIII
+4522.634, FeII
+4555.893, FeII
+4582.835, FeII
+4583.837, FeII
+4629.339, FeII
+4634.140, NIII
+4640.640, NIII
+4641.850, NIII
+4647.420, CIII
+4650.250, CIII
+4651.470, CIII
+4658.050, [FeIII]
+4685.710, HeII
+4711.260, [ArIV]
+4740.120, [ArIV]
+4861.333, Hb
+4893.370, [FeVII]
+4903.070, [FeIV]
+4923.927, FeII
+4958.911, [OIII]
+5006.843, [OIII]
+5018.440, FeII
+5084.770, [FeIII]
+5145.750, [FeVI]
+5158.890, [FeVII]
+5169.033, FeII
+5176.040, [FeVI]
+5197.577, FeII
+5200.257, [NI]
+5234.625, FeII
+5236.060, [FeIV]
+5270.400, [FeIII]
+5276.002, FeII
+5276.380, [FeVII]
+5302.860, [FeXIV]
+5309.110, [CaV]
+5316.615, FeII
+5316.784, FeII
+5335.180, [FeVI]
+5424.220, [FeVI]
+5517.709, [ClIII]
+5537.873, [ClIII]
+5637.600, [FeVI]
+5677.000, [FeVI]
+5695.920, CIII
+5720.700, [FeVII]
+5754.590, [NII]
+5801.330, CIV
+5811.980, CIV
+5875.624, HeI
+6046.440, OI
+6087.000, [FeVII]
+6300.304, [OI]
+6312.060, [SIII]
+6347.100, SiII
+6363.776, [OI]
+6369.462, FeII
+6374.510, [FeX]
+6516.081, FeII
+6548.050, [NII]
+6562.819, Ha
+6583.460, [NII]
+6716.440, [SII]
+6730.810, [SII]
+7002.230, OI
+7005.870, [ArV]
+7065.196, HeI
+7135.790, [ArIII]
+7155.157, [FeII]
+7170.620, [ArIV]
+7172.000, [FeII]
+7236.420, CII16.3
+7237.260, [ArIV]
+7254.448, OI
+7262.760, [ArIV]
+7281.349, HeI
+7319.990, [OII]
+7330.730, [OII]
+7377.830, [NiII]
+7411.160, [NiII]
+7452.538, [FeII]
+7468.310, NI
+7611.000, [SXII]
+7751.060, [ArIII]
+7816.136, HeI
+7868.194, ArI
+7889.900, [NiIII]
+7891.800, [FeXI]
+8236.790, HeII
+8392.397, Pa20
+8413.318, Pa19
+8437.956, Pa18
+8446.359, OI
+8467.254, Pa17
+8498.020, CaII
+8502.483, Pa16
+8542.090, CaII
+8545.383, Pa15
+8578.700, [ClII]
+8598.392, Pa14
+8616.950, [FeII]
+8662.140, CaII
+8665.019, Pa13
+8680.282, NI
+8703.247, NI
+8711.703, NI
+8750.472, Pa12
+8862.782, Pa11
+8891.910, [FeII]
+9014.909, Pa10
+9068.600, [SIII]
+9229.014, Pa9
+9531.100, [SIII]
+9545.969, Pae
+9824.130, [CI]
+9850.260, [CI]
+9913.000, [SVIII]
+10027.730, HeI
+10031.160, HeI
+10049.368, Pad
+10286.730, [SII]
+10320.490, [SII]
+10336.410, [SII]
+10746.800, [FeXIII]
+10830.340, HeI
+10938.086, Pag
